requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
psycopg2==2.9.3
pycodestyle==2.9.1
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.4
